fix(compliance): scope registry-seed heading stop to FR-shaped headings, not any heading
Doubt-review round 4b REJECT: the round-4 fix stopped the seeding scan at
ANY heading line to exclude a nested FR's own bullets, but
_iter_heading_blocks_by_index only treats an FR-shaped heading of
same-or-higher rank as a block boundary -- a non-FR heading (e.g. a plain
"#### Notes" subsection) interposed between two bullets of the SAME FR
block is not a boundary at all. Stopping there blinded seeding to a real,
already-marked bullet sitting past it: the exact round-3 failure mode
(a marker outside the seed pass's scan range lets a lost/empty registry
reissue its number), just triggered by an interposed heading instead of
an interposed prose paragraph -- and, unlike round 3's shape, this one
raised no exception at all, so the duplicate was permanently silent.

Fixed by scoping the stop condition to FR-shaped headings only (_HEADING_RE),
so seeding keeps scanning through a non-FR heading (like it already does
through prose) while still stopping at a genuine nested FR heading. Verified
against the new bug's fixture, the round-4 nested-FR fixture, and the round-3
prose fixture -- all three now behave correctly.

Also splits the growing golden-corpus tests into a third file
(test_ac_identity_seeding.py) to keep all three files under the 300-LOC
bloat-baseline threshold.

Run-ID: iterate-2026-09-06-p3-1-ac-identity-reader-corpus
