+++ .shipwright/planning/iterate/iterate-2026-09-06-p3-1-ac-identity-reader-corpus/reviews.json
@@ -168,12 +168,12 @@
     },
     "code": {
       "review_type": "code",
-      "status": "not_run",
+      "status": "completed",
       "findings_count": 0,
       "findings": [],
       "provider": null,
-      "completed_at": "2026-09-06T23:00:28.165073+00:00",
-      "disposition": "Stage-2 code-reviewer REJECTED at 3f-bis: HIGH finding, shared/scripts/lib/ac_identity.py:170 -- the registry seed pass reuses iter_bullet_positions, which is gated to the block's contiguous LEADING bullet run (_ac_blocks.py:134-144), so a [ACnn] marker sitting outside that run is invisible to seeding AND duplicate detection. Repro: an FR block with a bullet, an intervening prose line, then a bullet already carrying [AC01] -- with a lost/empty registry, mint() reseeds {} and stamps [AC01] onto the FIRST bullet too, producing two criteria under one FR sharing AC01 (the exact 'never reused' violation the seed pass exists to prevent), and the document becomes permanently un-mintable once the prose line is removed. Delivery stopped; PR #683 left open for repair (fix: split discovery into a leading-run scan for minting and an ungated whole-block scan for seeding/dup-check).",
+      "completed_at": "2026-09-07T05:15:43.911154+00:00",
+      "disposition": "Round 4 re-review after nested-heading regression fix (commit 0d04009a): PASS. Original round-3 HIGH (leading-run-gate seeding miss, fixed af19767d) and round-4 HIGH (nested-FR-heading false duplicate, fixed 0d04009a) both verified closed; no remaining findings.",
       "recorded_by": "code-reviewer",
       "parse_status": null,
       "raw_excerpt": null,
